Factor out api.py.
